Break logic into logical functions.
Created `deal`, `show_card`, `blind_moves` functions to separate logic.
Also added `button_or_blind` function to toss a coin and determine which
player is on the button first. This needs to be expanded, to alternate
button and blind positions, and to choose suitable method for tagging
players as being on the button or on the blind.